+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Prevent SonarCloud from running C-family analysis without compile commands
+- serve friend profile previews as HTML from Vercel
 
 ## [1.2.1] - 2026-04-13
 ### Fixed